feat: refactoring to use ESBUILD to bundle
