+++ src/decorators/resolver.decorator.test.ts
@@ -1,6 +1,6 @@
-import { Resolver } from '@/decorators'
-import { IoCContainer } from '@/container'
-import { ServiceScope } from '@/models'
+import { IoCContainer } from '../container/container'
+import { ServiceScope } from '../models/decorators'
+import { Resolver } from './resolver.decorator'
 
 describe('Resolver Decorator', () => {
   it('should define metadata and register the target in IoCContainer', () => {
