+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
   "compilerOptions": {
+    "module": "ESNext",
     "target": "ES2020",
-    "module": "CommonJS",
+    "moduleResolution": "Node",
     "esModuleInterop": true,
     "removeComments": true,
     "emitDecoratorMetadata": true,
@@ -9,11 +10,6 @@
     "experimentalDecorators": true,
     "allowSyntheticDefaultImports": true,
     "sourceMap": true,
-    "outDir": "./dist",
-    "baseUrl": "./",
-    "paths": {
-      "@/*": ["./src/*"]
-    },
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
@@ -21,12 +17,13 @@
     "strictBindCallApply": false,
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false,
-    "types": ["node", "vitest/globals"],
     "declaration": true,
-    "declarationDir": "./dist"
+    "declarationDir": "./dist",
+    "outDir": "./dist",
+    "types": ["node", "vitest/globals"]
   },
   "include": ["src"],
-  "exclude": ["node_modules", "*.test.ts"],
+  "exclude": ["node_modules", "*.test.ts", "dist"],
   "ts-node": {
     "require": ["tsconfig-paths/register"]
   }
